Introduces image_types_fwup bbclass
This commit adds support for fwup as IMAGE_TYPES. Now it's possible to
make fwup images automatically.

In order to use the bbclass it's necessary to add the following
configuration to your local.conf or similar:

    # Add image_types_fwup to the system
    IMAGE_CLASSES += "image_types_fwup"

    # Declare that we want to make fwup images
    IMAGE_FSTYPES = "fwup"

    # It's mandatory to define a fwup config file format
    FWUP_FILE = "sdimg-raspberrypi0-wifi.fwup"

That it's all. When running image creating with bitbake like:

    bitbake core-image-fullcmdline

A .fw file will be create as final result at tmp/deploy/images/<machine>/*.fw

That image is ready to burn using fwup:

    oe-run-native fwup-native fwup -a -d /dev/sda -t complete \
        -i build/tmp/deploy/images/raspberrypi0-wifi/core-image-full-cmdline-raspberrypi0-wifi.rootfs-20240920190231.fw

Note:

1) About the FWUP_FILE, the sdimg-raspberrypi0-wifi.fwup is just an example
   based on Nerver Project. As fwup is a pretty flexible tool you are in
   charge to inspect and create your own .fwup configuration file.

2) It's important to mention that the bbclass image_types_fwup adds the
   package 'fwup' to the target image.